🔧 FIX: Arduino serial communication and compilation issues
## 🎯 Key Improvements
- **Arduino Protocol Compatibility**: Fixed command format from uppercase to lowercase (M -> m) for Arduino compatibility
- **Native Windows API Migration**: Migrated from serial library to native Windows API for better performance and reliability
- **Auto-reconnection System**: Added intelligent connection health monitoring and automatic reconnection
- **CUDA Compilation**: Enabled previously disabled CUDA files in project configuration

## 🛠️ Technical Fixes
- **Compilation Warnings**: Resolved C4005 (macro redefinition), C4838 (narrowing conversion), C4819 (encoding), C2065 (undeclared identifier)
- **Serial Communication**: Implemented synchronous I/O with improved error handling and retry logic
- **Memory Management**: Enhanced connection state management with thread-safe operations
- **Project Configuration**: Updated vcxproj to include CUDA files and UTF-8 compiler flags

## 🚀 Performance Enhancements
- **Reduced Latency**: Eliminated unnecessary TIMER commands and debug logging
- **Improved Reliability**: Added connection validation and graceful degradation
- **Optimized Timeouts**: Configured appropriate timeouts for Arduino communication

## ✅ Validation
- Mouse movement functionality restored and working
- Recoil compensation system operational
- All compilation warnings resolved
- Arduino protocol compatibility confirmed

Fixes critical Arduino communication issues and eliminates all compilation warnings while maintaining functionality.

+++ needaimbot/capture/duplication_api_capture.cpp
@@ -321,7 +321,7 @@ public:
                  }
             }
 
-            UINT expectedDirtyRectsSize = (m_metaDataBuffer.size() >= moveRectsSize) ? (m_metaDataBuffer.size() - moveRectsSize) : 0;
+            UINT expectedDirtyRectsSize = (m_metaDataBuffer.size() >= moveRectsSize) ? static_cast<UINT>(m_metaDataBuffer.size() - moveRectsSize) : 0;
             UINT actualDirtyRectsSize = 0;
             if (expectedDirtyRectsSize > 0) {
                 BYTE* dirtyRectsDataStart = m_metaDataBuffer.data() + moveRectsSize; 
@@ -557,7 +557,7 @@ cv::cuda::GpuMat DuplicationAPIScreenCapture::GetNextFrameGpu()
                 LONG bottom = std::min(dirty.bottom, captureScreenRect.bottom);
                 if (left < right && top < bottom)
                 {
-                    D3D11_BOX box = { left, top, 0, right, bottom, 1 };
+                    D3D11_BOX box = { static_cast<UINT>(left), static_cast<UINT>(top), 0, static_cast<UINT>(right), static_cast<UINT>(bottom), 1 };
                     UINT destX = static_cast<UINT>(left - captureScreenRect.left);
                     UINT destY = static_cast<UINT>(top - captureScreenRect.top);
                     m_ddaManager->m_context->CopySubresourceRegion(
@@ -573,7 +573,7 @@ cv::cuda::GpuMat DuplicationAPIScreenCapture::GetNextFrameGpu()
         else
         {
             
-            D3D11_BOX box = { captureScreenRect.left, captureScreenRect.top, 0, captureScreenRect.right, captureScreenRect.bottom, 1 };
+            D3D11_BOX box = { static_cast<UINT>(captureScreenRect.left), static_cast<UINT>(captureScreenRect.top), 0, static_cast<UINT>(captureScreenRect.right), static_cast<UINT>(captureScreenRect.bottom), 1 };
             m_ddaManager->m_context->CopySubresourceRegion(
                 m_ddaManager->m_sharedTexture,
                 0,
@@ -649,7 +649,7 @@ cv::Mat DuplicationAPIScreenCapture::GetNextFrameCpu()
                 LONG bottom = std::min(dirty.bottom, captureScreenRect.bottom);
                 if (left < right && top < bottom)
                 {
-                    D3D11_BOX box = { left, top, 0, right, bottom, 1 };
+                    D3D11_BOX box = { static_cast<UINT>(left), static_cast<UINT>(top), 0, static_cast<UINT>(right), static_cast<UINT>(bottom), 1 };
                     UINT destX = static_cast<UINT>(left - captureScreenRect.left);
                     UINT destY = static_cast<UINT>(top - captureScreenRect.top);
                     m_ddaManager->m_context->CopySubresourceRegion(
@@ -663,7 +663,7 @@ cv::Mat DuplicationAPIScreenCapture::GetNextFrameCpu()
         else
         {
             
-            D3D11_BOX box = { captureScreenRect.left, captureScreenRect.top, 0, captureScreenRect.right, captureScreenRect.bottom, 1 };
+            D3D11_BOX box = { static_cast<UINT>(captureScreenRect.left), static_cast<UINT>(captureScreenRect.top), 0, static_cast<UINT>(captureScreenRect.right), static_cast<UINT>(captureScreenRect.bottom), 1 };
             m_ddaManager->m_context->CopySubresourceRegion(
                 m_ddaManager->m_sharedTexture, 0,
                 0, 0, 0,
